batch: Extract absence constraints
Batch merge still owns presence insertion, candidate enumeration, and
discard reversal orchestration.

Absence suppression has a separate lifecycle over realized entries: it
locates anchored boundaries, resolves deletion choices, and removes exact
forbidden sequences.

This commit moves absence-constraint application and ambiguity choices into
batch.absence_constraints, then rewires batch.merge to call that owner.

Architecture guards will pin the new absence-constraint boundary.
